requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# LLM & AI
openai>=1.3.0  # Still needed for embeddings
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.2
langchain-groq>=0.1.0
langchain-community>=0.0.10

# Multimodal processing
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0
easyocr>=1.7.0
openai-whisper>=20231117
pydub>=0.25.1

# Audio processing
soundfile>=0.12.1
librosa>=0.10.1

# Vector store & embeddings
faiss-cpu>=1.7.4
chromadb>=0.4.15
sentence-transformers>=2.2.2

# Math & computation
sympy>=1.12
numpy>=1.24.0
scipy>=1.11.0

# Utilities
pydantic>=2.5.0
tenacity>=8.2.3
tqdm>=4.66.0

# Data storage
json5>=0.9.14
